+++ Packages/RuuviDaemon/RuuviDaemon.podspec
@@ -30,6 +30,49 @@ Pod::Spec.new do |s|
     ss.dependency 'RuuviLocal'
   end
 
+  s.subspec 'Background' do |ss|
+    ss.source_files = 'Sources/RuuviDaemonBackground/**/*.{h,m,swift}', 'Sources/RuuviDaemonBackground/*.{h,m,swift}'
+    ss.dependency 'RuuviDaemon/Contract'
+    ss.dependency 'RuuviDaemon/Operation'
+  end
+
+  s.subspec 'Operation' do |ss|
+    ss.source_files = 'Sources/RuuviDaemonOperation/**/*.{h,m,swift}', 'Sources/RuuviDaemonOperation/*.{h,m,swift}'
+    ss.dependency 'RuuviDaemon/Contract'
+    ss.dependency 'RuuviStorage'
+    ss.dependency 'RuuviLocal'
+    ss.dependency 'RuuviPool'
+    ss.dependency 'RuuviVirtual'
+    ss.dependency 'RuuviOntology'
+    ss.dependency 'RuuviNotifier'
+    ss.dependency 'FutureX'
+  end
+
+  s.subspec 'RuuviTag' do |ss|
+    ss.source_files = 'Sources/RuuviDaemonRuuviTag/**/*.{h,m,swift}', 'Sources/RuuviDaemonRuuviTag/*.{h,m,swift}'
+    ss.dependency 'RuuviDaemon/Contract'
+    ss.dependency 'BTKit'
+    ss.dependency 'RuuviLocal'
+    ss.dependency 'RuuviPool'
+    ss.dependency 'RuuviReactor'
+    ss.dependency 'RuuviStorage'
+    ss.dependency 'RuuviPersistence'
+    ss.dependency 'RuuviOntology'
+    ss.dependency 'RuuviService'
+    ss.dependency 'RuuviNotification'
+    ss.dependency 'RuuviNotifier'
+  end
+
+  s.subspec 'VirtualTag' do |ss|
+    ss.source_files = 'Sources/RuuviDaemonVirtualTag/**/*.{h,m,swift}', 'Sources/RuuviDaemonVirtualTag/*.{h,m,swift}'
+    ss.dependency 'RuuviDaemon/Contract'
+    ss.dependency 'BTKit'
+    ss.dependency 'RuuviLocal'
+    ss.dependency 'RuuviOntology'
+    ss.dependency 'RuuviVirtual'
+    ss.dependency 'RuuviNotifier'
+  end
+
   s.test_spec 'Tests' do |test_spec|
     test_spec.source_files = 'Tests/**/*.{swift}', 'Tests/*.{swift}'
   end
