+++ Packages/RuuviDaemon/Sources/RuuviDaemon/BackgroundProcessService.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public protocol BackgroundProcessService {
+    func register()
+    func schedule()
+}
