+++ Packages/RuuviDaemon/Sources/RuuviDaemon/BackgroundTaskService.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public protocol BackgroundTaskService {
+    func register()
+    func schedule()
+}
