+++ Packages/RuuviDaemon/Sources/RuuviDaemon/PullWebDaemon.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public protocol PullWebDaemon {
+    func start()
+    func stop()
+    func wakeUp()
+}
