+++ Packages/RuuviDaemon/Sources/RuuviDaemon/RuuviTagAdvertisementDaemon.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+extension Notification.Name {
+    public static let RuuviTagAdvertisementDaemonDidFail = Notification.Name("RuuviTagAdvertisementDaemonDidFail")
+}
+
+public enum RuuviTagAdvertisementDaemonDidFailKey: String {
+    case error = "RuuviDaemonError" // RuuviDaemonError
+}
+
+public protocol RuuviTagAdvertisementDaemon {
+    func start()
+    func stop()
+}
