+++ Packages/RuuviDaemon/Sources/RuuviDaemon/RuuviTagPropertiesDaemon.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+extension Notification.Name {
+    public static let RuuviTagPropertiesDaemonDidFail = Notification.Name("RuuviTagPropertiesDaemonDidFail")
+}
+
+public enum RuuviTagPropertiesDaemonDidFailKey: String {
+    case error = "RuuviDaemonError" // RuuviDaemonError
+}
+
+public protocol RuuviTagPropertiesDaemon {
+    func start()
+    func stop()
+}
