--- station/Classes/Business/Background/Process/BackgroundProcessService.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-protocol BackgroundProcessService {
-    func register()
-    func schedule()
-}
