--- station/Classes/Business/Background/Task/BackgroundTaskService.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-protocol BackgroundTaskService {
-    func register()
-    func schedule()
-}
