--- station/Classes/Business/Daemons/Background/PullWeb/PullWebDaemon.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-protocol PullWebDaemon {
-    func start()
-    func stop()
-    func wakeUp()
-}
