--- station/Classes/Business/Daemons/Foreground/Advertisement/RuuviTagAdvertisementDaemon.swift
@@ -1,14 +1,0 @@
-import Foundation
-
-extension Notification.Name {
-    static let RuuviTagAdvertisementDaemonDidFail = Notification.Name("RuuviTagAdvertisementDaemonDidFail")
-}
-
-enum RuuviTagAdvertisementDaemonDidFailKey: String {
-    case error = "RUError" // RUError
-}
-
-protocol RuuviTagAdvertisementDaemon {
-    func start()
-    func stop()
-}
