--- station/Classes/Business/Daemons/Foreground/Properties/RuuviTagPropertiesDaemon.swift
@@ -1,14 +1,0 @@
-import Foundation
-
-extension Notification.Name {
-    static let RuuviTagPropertiesDaemonDidFail = Notification.Name("RuuviTagPropertiesDaemonDidFail")
-}
-
-enum RuuviTagPropertiesDaemonDidFailKey: String {
-    case error = "RUError" // RUError
-}
-
-protocol RuuviTagPropertiesDaemon {
-    func start()
-    func stop()
-}
