--- station/Classes/Business/Daemons/Foreground/WebTag/VirtualTagDaemon.swift
@@ -1,14 +1,0 @@
-import Foundation
-
-extension Notification.Name {
-    static let WebTagDaemonDidFail = Notification.Name("WebTagDaemonDidFail")
-}
-
-enum WebTagDaemonDidFailKey: String {
-    case error = "RUError" // RUError
-}
-
-protocol VirtualTagDaemon {
-    func start()
-    func stop()
-}
